+++ test/test-uv-mapping.c
@@ -1,6 +1,6 @@
 #include <assert.h>
 #include <stdlib.h>
-#include "uv_mapping.h"
+#include "../src/uv_mapping.h"
 
 #define CHECK_ERR(err) assert(uvwasi__translate_uv_error(UV_ ## err ) ==      \
                               UVWASI_ ## err)
